Fix errors with Click testing

+++ bot/app/data/models/payments.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm import relationship
 class Payment(Base):
     __tablename__ = 'payments'
 
-    id = sa.Column(sa.Integer, primary_key=True)
+    id = sa.Column(sa.BigInteger, primary_key=True)
     amount = sa.Column(sa.Integer)
     provider = sa.Column(sa.String)
     user_id = sa.Column(sa.Integer, sa.ForeignKey('users.id'))
